+++ packages/cortex-react/docs/ProgressBar.stories.tsx
@@ -1,4 +1,4 @@
-import { Meta, StoryObj } from '@storybook/react';
+import { Meta, StoryObj } from '@storybook/react-vite';
 import { ProgressBar } from '../src';
 
 export default {