+++ packages/rn-playground/stories/Button/Button.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-vite';
 import { Button, Text } from '@tecsinapse/react-native-kit';
 
 const ButtonMeta: Meta<typeof Button> = {