+++ packages/rn-playground/stories/GroupButton/GroupButton.stories.tsx
@@ -4,7 +4,7 @@ import {
   GroupButtonValue,
 } from '@tecsinapse/react-native-kit';
 import React, { useState } from 'react';
-import { Meta } from '@storybook/react';
+import { Meta } from '@storybook/react-vite';
 
 const StoryMeta: Meta<typeof GroupButton> = {
   title: 'GroupButton',