+++ packages/rn-playground/stories/PhoneInput/PhoneInput.stories.tsx
@@ -1,4 +1,4 @@
-import { Meta, StoryFn } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react-vite';
 import { PhoneInput } from '@tecsinapse/react-native-kit';
 import React, { useState } from 'react';
 import { View } from 'react-native';